func_8018CA94 + file split
