+++ config/splat.us.bomar.yaml
@@ -53,8 +53,9 @@ segments:
     - [0xA740, rodata]
     - [0xA8A0, rodata]
     - [0xA910, rodata]
-    - [0xAB10, rodata]
-    - [0xABD0, .rodata, AC0C] # func_80198688
+    - [0xAA70, .rodata, AC0C]
+    - [0xAAB0, rodata]
+    - [0xABD0, .rodata, CF70] # func_80198688
     - [0xABE8, rodata]
     - [0xAC0C, c]
     - [0x1A200, sbss]
