+++ src/boss/mar/AC0C.c
@@ -24,762 +24,196 @@ INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018B74C);
 
 INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018B850);
 
+// https://decomp.me/scratch/qHybN Matching,
+// on hold until D_80097400 type descovery is resolved
 INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018C90C);
 
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018CA94);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018CF70);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018CFA0);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018D29C);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018D3A4);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018E4BC);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018EB60);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018EC24);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018ED3C);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018EE54);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018EEA0);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018EEF8);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018EFF4);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018F108);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018F154);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018F1AC);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018F2A8);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018F3BC);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018F534);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", CreateEntityFromCurrentEntity);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018F65C);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018F6D8);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018F750);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80190388);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_801903F4);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80190470);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_801904B8);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80190570);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80190688);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_801906C4);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_801906F8);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8019073C);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8019076C);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80190798);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80190A10);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80190BF8);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80190C58);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80190C84);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80190CA0);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80190D0C);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80190D44);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80190D8C);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80190DD4);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80190E2C);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80190EB8);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80190EE8);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80190F20);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80190F68);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80190FC8);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80190FE8);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80191004);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", InitializeEntity);
-
-void EntityDummy(Entity* self) {
-    if (self->step == 0) {
-        self->step++;
-    }
-}
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_801911BC);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_801912E4);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8019143C);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80191780);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80191838);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_801918BC);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80191A18);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80191A98);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80191B74);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80191C8C);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80191D30);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80191D80);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80191DA8);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8019261C);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80192718);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_801927BC);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80192DA4);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80193054);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80193B24);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80193C40);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8019419C);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8019428C);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8019439C);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_801944CC);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_801945B4);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_801946A4);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80194798);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_801949C8);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80194AC4);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80195044);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80195144);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80195210);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_801952D8);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80195390);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80195568);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80195B48);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80196248);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80196440);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80196538);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80196708);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8019738C);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80197720);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80197BD0);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80197CBC);
-
-// ID 0x16
-void EntityClockRoomController(Entity* self) {
-    PlayerStatus* status = &g_Status;
-    const int birdcageDoor1 = 7;
-    const int birdcageDoor2 = 8;
-    const int stoneDoors = 14;
-    const int minuteHand = 5;
-    const int hourHand = 6;
-    const int statues = 1;
-    const int shadow = 9;
-    const int gears = 12;
-    Entity* newEntity;
-    Primitive* prim;
-    s16 primIndex;
-    u16 i, j;
-
-    // Plays the clock bell
-    if (self->ext.clockRoom.bellTimer != 0) {
-        if (self->ext.clockRoom.bellDuration == 0) {
-            g_api.PlaySfx(0x7A6); // Clock bell
-            if (--self->ext.clockRoom.bellTimer) {
-                self->ext.clockRoom.bellDuration = 64;
-            }
-        } else {
-            self->ext.clockRoom.bellDuration--;
-        }
-    }
-
-    // Controls the statues
-    if (D_800973FC == 0) {
-        if (PLAYER.posY.i.hi > 128) {
-            g_Statues[0] = false; // right statue closed
-        }
-    } else if (self->ext.clockRoom.unk8A == 0) {
-        g_Statues[0] = true; // right statue open
-    }
-
-    self->ext.clockRoom.unk8A = D_800973FC;
-    if (self->step != 0) {
-        if ((status->timerFrames == 0) && (status->timerSeconds == 0)) {
-            if (status->timerMinutes & 1) {
-                g_Statues[1] = false; // left statue closed
-            } else {
-                g_Statues[1] = true; // left statue open
-            }
-        }
-    } else if (status->timerMinutes & 1) {
-        g_Statues[1] = false; // left statue closed
-    } else {
-        g_Statues[1] = true; // left statue open
+// INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8018CA94);
+
+extern u8 D_80180690[];
+extern u8 D_801806A4[];
+extern u8 D_801806B8[];
+extern u8 D_801806CC[];
+extern u8 D_801806E4[];
+extern u8 D_801806EC[];
+extern u8 D_80180700[];
+extern u8 D_80180718[];
+extern u8 D_80180724[];
+extern u8 D_80180738[];
+extern s32 D_8019AE70;
+extern s32 D_8019AF20;
+
+void func_8018CA94(Entity* self) {
+    s32 posX;
+    s32 posX2;
+    s32 temp;
+    s32 var_a2_2;
+    s32 var_v0;
+    s32 var_v0_2;
+    s32 var_v0_3;
+    u16 temp_v1;
+    u16 temp_v1_2;
+    u16 temp_v1_4;
+
+    if ((D_8019AE70 != 0) && (self->step < 12)) {
+        func_80190FC8(12); // SetStep
     }
 
     switch (self->step) {
-    case 0:
-        if (g_blinkTimer == ((g_blinkTimer / 60) * 0x3C)) {
-            g_api.PlaySfx(0x7A9); // Clock tick
-        }
+        case 0:
+            InitializeEntity(D_801803B4);
+            self->animSet = -0x7FFE;
+            self->animCurFrame = 1;
+            self->unk5A = 0x48;
+            self->palette = 0x210;
+            break;
 
-        primIndex = g_api.AllocPrimitives(PRIM_G4, 1);
-        if (primIndex == -1) {
-            return;
-        }
-        self->primIndex = primIndex;
-        self->flags |= FLAG_HAS_PRIMS;
-        prim = &g_PrimBuf[primIndex];
-        prim->r0 = prim->g0 = prim->b0 = 0;
-        prim->y2 = prim->y3 = prim->x1 = prim->x3 = 0x100;
-        prim->priority = 0x1F0;
-        prim->y0 = prim->y1 = prim->x0 = prim->x2 = 0;
-        prim->blendMode = BLEND_VISIBLE;
-        LOW(prim->r1) = LOW(prim->r0);
-        LOW(prim->r2) = LOW(prim->r0);
-        LOW(prim->r3) = LOW(prim->r0);
-
-        InitializeEntity(D_801803E4);
-        g_api.PlaySfx(SET_STOP_MUSIC);
-        D_80097928 = 1;
-        g_Statues[0] = false; // right statue closed
-        self->animSet = ANIMSET_OVL(1);
-        self->animCurFrame = 23;
-        D_80097910 = 0;
-        self->zPriority = 0x40;
-
-        // Create clock hands
-        newEntity = &self[5];
-        for (i = 0; i < 2; i++) {
-            CreateEntityFromCurrentEntity(E_CLOCK_HANDS, newEntity);
-            newEntity->params = i;
-            newEntity++;
-        }
-
-        // Clock set
-        self[minuteHand].ext.clockRoom.hand = status->timerMinutes * 0x3C;
-        self[hourHand].ext.clockRoom.hand =
-            (status->timerHours * 0x12C) + (status->timerMinutes * 5);
-
-        // Create Birdcage doors
-        newEntity = &self[birdcageDoor1];
-        for (i = 0; i < 2; i++) {
-            CreateEntityFromCurrentEntity(E_BIRDCAGE_DOOR, newEntity);
-            newEntity->params = i;
-            newEntity++;
-        }
-
-        if ((status->timerMinutes >= 10) && (status->timerMinutes < 30)) {
-            self[birdcageDoor1].ext.birdcage.state = true;
-        } else {
-            self[birdcageDoor1].ext.birdcage.state = false;
-        }
-
-        if ((status->timerMinutes >= 30) && (status->timerMinutes < 50)) {
-            self[birdcageDoor2].ext.birdcage.state = true;
-        } else {
-            self[birdcageDoor2].ext.birdcage.state = false;
-        }
-
-        // Shadow for the Bighorn sheep head on the center
-        CreateEntityFromCurrentEntity(E_DUMMY_1D, &self[shadow]);
-        self[shadow].animSet = ANIMSET_OVL(1);
-        self[shadow].animCurFrame = 23;
-        self[shadow].zPriority = 0x40;
-        self[shadow].palette = 0x804B;
-        self[shadow].unk19 = 8;
-        self[shadow].blendMode = 0x10;
-        self[shadow].flags = FLAG_DESTROY_IF_OUT_OF_CAMERA | FLAG_UNK_08000000 |
-                             FLAG_UNK_04000000;
-        self[shadow].posY.i.hi += 4;
-
-        // Create path blocking statues
-        newEntity = &self[statues];
-        for (i = 0; i < 2; i++) {
-            CreateEntityFromCurrentEntity(E_STATUE, newEntity);
-            newEntity->params = i;
-            newEntity++;
-        }
-
-        // Create the gears that drive the statues
-        newEntity = &self[gears];
-        for (j = 0; j < 2; j++) {
-            CreateEntityFromCurrentEntity(E_STATUE_GEAR, newEntity);
-            newEntity->params = j;
-            newEntity++;
-        }
-
-        // Create the stones on the floor
-        newEntity = &self[stoneDoors];
-        for (j = 0; j < 2; j++) {
-            CreateEntityFromCurrentEntity(E_STONE_DOOR, newEntity);
-            newEntity->params = j;
-            newEntity++;
-        }
-        break;
-
-    case 1:
-        if (status->timerFrames == 0) {
-            g_api.PlaySfx(0x7A9); // clock tick
-        }
-
-        // Update clock hands
-        self[minuteHand].ext.clockRoom.hand = status->timerMinutes * 0x3C;
-        self[hourHand].ext.clockRoom.hand =
-            (status->timerHours * 0x12C) + (status->timerMinutes * 5);
-        if ((status->timerSeconds == 0) && (status->timerFrames == 0) &&
-            (status->timerMinutes == 0)) {
-            self->ext.clockRoom.bellTimer =
-                ((status->timerHours + 11) % 12) + 1;
-            if (self->ext.clockRoom.bellTimer == 0) {
-                self->ext.clockRoom.bellTimer = 12;
-            }
-        }
-        break;
-    }
-}
-
-// Entity ID 0x17
-void EntityClockHands(Entity* self) {
-    Entity* handShadow = &self[5];
-    u16 params = self->params;
-
-    if (self->step == 0) {
-        InitializeEntity(D_801803E4);
-        self->animSet = ANIMSET_OVL(1);
-        self->animCurFrame = params + 25;
-        self->zPriority = 0x3F - params;
-        self->unk19 = 4;
-
-        // Create hand shadows
-        CreateEntityFromCurrentEntity(E_DUMMY_1D, handShadow);
-        handShadow->unk19 = 0xC;
-        handShadow->blendMode = 0x10;
-        handShadow->animSet = ANIMSET_OVL(1);
-        handShadow->animCurFrame = params + 25;
-        handShadow->zPriority = 0x3F - params;
-        handShadow->flags = FLAG_DESTROY_IF_OUT_OF_CAMERA | FLAG_UNK_08000000 |
-                            FLAG_UNK_04000000;
-        handShadow->posY.i.hi += 4;
-    }
-
-    self->rotAngle = (self->ext.clockRoom.hand * 0x1000) / 3600;
-    if (params != 0) {
-        self->rotAngle += 0x400;
-    }
-
-    handShadow->rotAngle = self->rotAngle &= 0xFFF;
-}
-
-// Birdcage doors on the clock Entity ID 0x18
-void EntityBirdcageDoor(Entity* self) {
-    u16 params = self->params;
-
-    switch (self->step) {
-    case 0:
-        InitializeEntity(D_801803E4);
-        self->animSet = ANIMSET_OVL(1);
-        self->animCurFrame = D_801812A8[self->ext.birdcage.state & 1];
-        self->zPriority = 0x3C;
-        self->unk1A = self->unk1C = 0x100;
-        self->ext.birdcage.prevState = self->ext.birdcage.state;
-        self->unk6C = 0x80;
-        self->posX.i.hi = D_80181280[params] - g_Camera.posX.i.hi;
-        self->posY.i.hi = D_80181284[params] - g_Camera.posY.i.hi;
-        break;
-
-    case 1:
-        if (self->ext.birdcage.prevState != self->ext.birdcage.state) {
-            self->unk19 = 0xB;
-            self->ext.birdcage.timer = 64;
-            self->ext.birdcage.prevState = self->ext.birdcage.state;
-            self->step++;
-            g_api.PlaySfx(0x608);
-        }
-        break;
-
-    case 2:
-        self->unk1A = self->unk1C -= 2;
-        self->unk6C += 0xFF;
-        if (--self->ext.birdcage.timer == 0) {
-            self->ext.birdcage.timer = 64;
-            self->zPriority = 0;
-            self->step++;
-            g_api.PlaySfx(0x608);
-        }
-        break;
-
-    case 3:
-        self->posX.val += 0x2000;
-        if (--self->ext.birdcage.timer == 0) {
-            self->ext.birdcage.timer = 64;
-            self->animCurFrame = D_801812A8[self->ext.birdcage.state & 1];
-            self->posX.i.hi -= 8;
-            self->posY.i.hi += 8;
-            self->step++;
-            g_api.PlaySfx(0x608);
-        }
-        break;
-
-    case 4:
-        self->posY.val -= 0x2000;
-        if (--self->ext.birdcage.timer == 0) {
-            self->ext.birdcage.timer = 64;
-            self->zPriority = 0x3C;
-            self->step++;
-            g_api.PlaySfx(0x608);
-        }
-        break;
-
-    case 5:
-        self->unk1A = self->unk1C += 2;
-        self->unk6C += 1;
-        if (--self->ext.birdcage.timer == 0) {
-            self->unk19 = 0;
-            self->step = 1;
-        }
-        break;
-    }
-}
-
-void UpdateStatueTiles(s32 tilePos, s32 tile) {
-    u32 i;
-
-    for (i = 0; i < 6; i++) {
-        g_CurrentRoomTileLayout.fg[tilePos] = tile;
-        tilePos++;
-        g_CurrentRoomTileLayout.fg[tilePos] = tile;
-        tilePos += 15;
-    }
-}
-
-// Entity ID 0x19
-void EntityStatue(Entity* self) {
-    Entity* entity = &self[2];
-    Entity* statueGear = &self[11];
-    u16 params = self->params;
-    u16* temp_a0;
-    u16* fakeVar;
-
-    switch (self->step) {
-    case 0:
-        InitializeEntity(D_801803E4);
-        self->animSet = ANIMSET_OVL(1);
-        self->animCurFrame = params + 10;
-        self->hitboxWidth = 16;
-        self->hitboxHeight = 32;
-        self->zPriority = 0x40;
-
-        if (g_Statues[params] == 0) {
-            self->posX.i.hi += D_80181288[params];
-            if (self->params != 0) {
-                UpdateStatueTiles(2, 0x597);
-            } else {
-                UpdateStatueTiles(12, 0x597);
-            }
-        } else {
-            self->posX.i.hi += D_8018128C[params];
-            if (self->params != 0) {
-                UpdateStatueTiles(2, 0);
-            } else {
-                UpdateStatueTiles(12, 0);
-            }
-        }
-
-        self->ext.statue.step = g_Statues[params];
-        self->posY.i.hi -= 58;
-
-        // Create shadow for the statue
-        CreateEntityFromCurrentEntity(E_DUMMY_1D, entity);
-        entity->animSet = ANIMSET_OVL(1);
-        entity->animCurFrame = params + 10;
-        entity->zPriority = 0x3F;
-        entity->unk19 = 8;
-        entity->blendMode = 0x10;
-        entity->flags = FLAG_DESTROY_IF_OUT_OF_CAMERA | FLAG_UNK_08000000 |
-                        FLAG_UNK_04000000;
-        entity->posY.i.hi += 8;
-        break;
-
-    case 1:
-        //! FAKE
-        fakeVar = g_Statues;
-        temp_a0 = fakeVar;
-        temp_a0 += params;
-        if (*temp_a0 != self->ext.statue.step) {
-            self->ext.statue.step = *temp_a0;
-            if (self->ext.statue.step == 0) {
-                statueGear->ext.statue.step = 2;
-            } else {
-                statueGear->ext.statue.step = 1;
-            }
-            self->hitboxState = 2;
-            self->step++;
-            func_80196440(0x609);
-        }
-        break;
-
-    case 2:
-        func_8019143C(self, 0x10, 0x20, 0x13);
-        if (self->step_s == 0) {
-            if (self->ext.statue.step != 0) {
-                if (self->params != 0) {
-                    UpdateStatueTiles(2, 0);
-                } else {
-                    UpdateStatueTiles(12, 0);
-                }
-            }
-            self->ext.statue.timer = 96;
-            self->step_s++;
-        }
-
-        if (self->ext.statue.step != 0) {
-            self->posX.val += D_801812A0[params];
-        } else {
-            self->posX.val -= D_801812A0[params];
-        }
-
-        if (--self->ext.statue.timer == 0) {
-            if (self->ext.statue.step == 0) {
-                if (self->params != 0) {
-                    UpdateStatueTiles(2, 0x597);
-                } else {
-                    UpdateStatueTiles(12, 0x597);
-                }
-            }
-            statueGear->ext.statue.step = 0;
-            self->hitboxState = 0;
-            self->step_s = 0;
-            self->step--;
-        }
-        break;
-    }
-    entity->posX.i.hi = self->posX.i.hi;
-}
-
-// Gears that spin while the statues are moving Entity ID 0x1A
-void EntityStatueGear(Entity* self) {
-    u16 params = self->params;
-    Primitive* prim;
-    s16 primIndex;
-
-    switch (self->step) {
-    case 0:
-        if (self->step_s == 0) {
-            InitializeEntity(D_801803E4);
-            self->animSet = ANIMSET_OVL(1);
-            self->animCurFrame = 17;
-            self->zPriority = 0x80;
-            self->posX.i.hi += D_80181290[params];
-            self->posY.i.hi -= 44;
-            self->step = 0;
-            self->step_s++;
-        }
-
-        primIndex = g_api.AllocPrimitives(PRIM_TILE, 1);
-        if (primIndex == -1) {
-            return;
-        }
-        self->primIndex = primIndex;
-        self->flags |= FLAG_HAS_PRIMS;
-        prim = &g_PrimBuf[primIndex];
-        prim->r0 = prim->g0 = prim->b0 = 0;
-        prim->x0 = self->posX.i.hi - 6;
-        prim->y0 = self->posY.i.hi - 16;
-        prim->u0 = 12;
-        prim->v0 = 32;
-        prim->priority = 0x7F;
-        prim->blendMode = 0;
-        self->step++;
-        break;
-
-    case 1:
-        if (self->ext.statue.step == self->step) {
-            self->step = 2;
-            self->animFrameIdx = self->animCurFrame - 17;
-        }
-        if (self->ext.statue.step == 2) {
-            self->step = 3;
-            self->animFrameIdx = 20 - self->animCurFrame;
-        }
-        self->animFrameDuration = 0;
-        break;
-
-    case 2:
-        func_801904B8(D_801812C4, self);
-        if (self->ext.statue.step == 0) {
-            self->step = 1;
-        }
-        break;
-
-    case 3:
-        func_801904B8(D_801812D0, self);
-        if (self->ext.statue.step == 0) {
-            self->step = 1;
-        }
-        break;
-    }
-}
-
-void UpdateStoneDoorTiles(bool doorState) {
-    s32 tilePos;
-    s16 i, j;
-
-    for (tilePos = 0xC4, i = 0; i < 2; i++) {
-        for (j = 0; j < 8; j++) {
-            if (doorState) {
-                // Open stone doors
-                g_CurrentRoomTileLayout.fg[tilePos] = 0x597;
-                tilePos++;
-            } else {
-                // Close stone doors
-                g_CurrentRoomTileLayout.fg[tilePos] = g_StoneDoorTiles[j];
-                tilePos++;
-            }
-        }
-        tilePos += 8;
-    }
-}
-
-// Stone doors on the floor leading to CEN Entity ID 0x1B
-void EntityStoneDoor(Entity* self) {
-    u16 params = self->params;
-    const int centerCubeDoor = 0;
-
-    switch (self->step) {
-    case 0:
-        InitializeEntity(D_801803E4);
-        self->animSet = ANIMSET_OVL(1);
-        self->animCurFrame = params + 27;
-        self->zPriority = 0x40;
-        if (D_8003BDEC[centerCubeDoor] == 0) {
-            self->posX.i.hi += D_80181294[params];
-            UpdateStoneDoorTiles(true);
-        } else {
-            self->posX.i.hi += D_80181298[params];
-            UpdateStoneDoorTiles(false);
-        }
-        self->posY.i.hi += 88;
-        self->ext.stoneDoor.flag = D_8003BDEC[centerCubeDoor];
-        break;
-
-    case 1:
-        if (self->ext.stoneDoor.flag == NULL) {
-            if (D_8003BDEC[centerCubeDoor]) {
-                self->ext.stoneDoor.unk80 = 0;
+        case 1:
+            if (D_8019AF20 & 4) {
+                self->animCurFrame = 2;
                 self->step++;
             }
-        }
-        self->ext.stoneDoor.flag = D_8003BDEC[centerCubeDoor];
-        break;
+            break;
 
-    case 2:
-        if ((self->ext.stoneDoor.unk80 % 32) == 0) {
-            func_80196440(0x607);
-        }
-
-        if (++self->ext.stoneDoor.unk80 % 2) {
-            if (params != 0) {
-                self->posX.i.hi++;
-            } else {
-                self->posX.i.hi--;
+        case 2:
+            if (D_8019AF20 & 8) {
+                func_80190FC8(3);
             }
-            if (self->ext.stoneDoor.unk80 % 2) {
-                g_backbufferY = 1;
-            } else {
-                g_backbufferY = 0;
-            }
-        } else {
-            g_backbufferY = 0;
-        }
+            break;
 
-        if (self->ext.stoneDoor.unk80 > 96) {
-            UpdateStoneDoorTiles(false);
-            g_backbufferY = 0;
-            self->step--;
-        }
-        break;
+        case 3:
+            func_801904B8(D_80180690, self);
+            if (D_8019AF20 & 0x10) {
+                func_80190FC8(4);
+            }
+            break;
+
+        case 4:
+            func_801904B8(D_801806A4, self);
+            if (D_8019AF20 & 0x20) {
+                func_80190FC8(5);
+                return;
+            }
+            break;
+
+        case 5:
+            func_801904B8(D_801806B8, self);
+            if (D_8019AF20 & 0x40) {
+                func_80190FC8(6);
+            }
+            break;
+
+        case 6:
+            self->animCurFrame = 19;
+            if (D_8019AF20 & 0x80) {
+                self->step++;
+            }
+            break;
+
+        case 7:
+            self->animCurFrame = 0x12;
+            if (D_8019AF20 & 0x100) {
+                self->step++;
+            }
+            break;
+
+        case 8:
+            func_801904B8(D_801806CC, self);
+            if (D_8019AF20 & 0x200) {
+                func_80190FC8(9);
+            }
+            break;
+
+        case 9:
+            func_801904B8(D_801806E4, self);
+            if (D_8019AF20 & 0x400) {
+                func_80190FC8(10);
+            }
+            break;
+
+        case 10:
+            func_801904B8(D_801806EC, self);
+            if (D_8019AF20 & 0x800) {
+                func_80190FC8(11);
+            }
+            break;
+
+        case 11:
+            func_801904B8(D_80180700, self);
+            if (D_8019AF20 & 0x1000) {
+                func_80190FC8(12);
+            }
+            break;
+
+        case 12:
+            if (func_801904B8(D_80180718, self) == 0) {
+                func_80190FC8(13);
+                self->velocityX = 0x18000;
+            }
+            break;
+
+        case 13:
+            if ((func_801904B8(D_80180724, self) & 0x80) && ((self->animFrameIdx == 3) || (self->animFrameIdx == 7))) {
+                posX = self->posX.i.hi - 120;
+                temp = posX >> 4;
+
+                if (posX < 0) {
+                    temp = (self->posX.i.hi - 105) >> 4;
+                }
+                if (temp < -8) {
+                    temp = -8;
+                }
+                if (temp > 8) {
+                    temp = 8;
+                }
+                g_api.func_80134714(0x64C, 0x50, temp);
+            }
+
+            MoveEntity();
+            if (self->posX.i.hi >= 0xB9) {
+                func_80190FC8(0xE);
+                self->velocityY = -0x40000;
+                return;
+            }
+            break;
+
+        case 14:
+            func_801904B8(D_80180738, self);
+            self->velocityY += 0x3000;
+            MoveEntity();
+            if ((self->velocityY > 0) && (self->posY.i.hi >= 132)) {
+                self->velocityY = 0;
+                func_80190FC8(15);
+            }
+            break;
+
+        case 15:
+            if ((func_801904B8(D_80180724, self) & 0x80) && ((self->animFrameIdx == 3) || (self->animFrameIdx == 7))) {
+                posX2 = self->posX.i.hi - 120;
+                temp = posX2 >> 4;
+
+                if (posX2 < 0) {
+                    temp = (self->posX.i.hi - 105) >> 4;
+                }
+                if (temp < -8) {
+                    temp = -8;
+                }
+                if (temp > 8) {
+                    temp = 8;
+                }
+                g_api.func_80134714(0x64C, 0x50, temp);
+            }
+
+            MoveEntity();
+            if (self->posX.i.hi > 256) {
+                D_8019AF20 |= 0x2000;
+            }
+            if (self->posX.i.hi > 272) {
+                func_80190388(self);
+            }
+            break;
     }
 }
 
-// Entity ID 0x1C
-void func_80199114(Entity* self) {}
 
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8019911C);
 
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80199504);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80199654);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80199684);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80199700);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80199788);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_80199868);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_801998BC);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_801998E8);
-
-INCLUDE_ASM("asm/us/boss/mar/nonmatchings/AC0C", func_8019994C);
